requirements.txt
dlib==19.24.2
fastapi==0.100.0
ftfy==6.1.1
ipython==8.14.0
numexpr==2.8.4
numpy==1.24.3
opencv_python==4.6.0.66
opencv_python_headless==4.8.0.74
pandas==2.0.3
Pillow==9.2.0
Pillow==10.0.0
regex==2022.9.13
rembg==2.0.49
setuptools==68.0.0
skimage==0.0
tensorflow==2.10.0
tensorflow_addons==0.17.1
torch==2.0.1
tqdm==4.64.1
uvicorn==0.22.0
